requirements.txt
# KLARIQO DEPENDENCIES
# Core framework
flask==2.3.3
flask-sock==0.7.0

# Environment management
python-dotenv==1.0.0

# AI & Voice Services
groq==0.4.2
openai==1.35.3
elevenlabs==0.2.24
deepgram-sdk==3.2.7
google-generativeai==0.3.2

# Telephony services
twilio==8.9.1

# Data processing
pandas==2.1.3
requests==2.31.0

# Audio processing (CRITICAL for production)
audioop-lts==0.2.1  # For Python 3.12+ compatibility
librosa==0.10.1     # MP3 to PCM conversion for TTS fallback
numpy==1.24.3       # Required by librosa for audio processing

# Optional: Google Sheets integration (future)
# google-auth==2.23.4
# google-auth-oauthlib==1.1.0
# google-auth-httplib2==0.1.1
# google-api-python-client==2.108.0